I fixed the problem raised by 45acp.  If the user sets firefox to open their homepage when it is started, and the homepage is the same URL as browser.newtab then the focus will be placed correctly.  Also, it seems the bug raised by some others (url bar getting focus when a new tab finishes loading) is not happening.
The last problem is this: If the user opens a new tab via middle click or using the context menu on some link, the focus will be changed by my addon.  This is not the desired behavior but
I have no way to differentiate between tabs opened these two ways or via the green plus.  I can't check the URL because some sites change it (e.g., the user may set their URL as
slashdot.org but this may be redirected to beta.slashdot.org)