Fixed the 'new window' bug.
Going forward in the future.  I should consider handling things in two cases: 1) New Tabs 2) New Windows.  And, when a new window is open, there is a new tab here (that is right now a
special case).  The new window function should call the new tab function on it's own initial tab, and then set up the NewTab event listener in the window for whatever else tabs the user
creates.  I think though, that the current implementation was very easy to add support for the first tab in a new window.  Maybe it's fine as is.